[zstd][errors] Get rid of static error code translation and use C call
- Errors were defined in private C/header files and we needed to change them every release (FIXME: this is very fragile, must map 1-to-1 with zstd's)
- We actually just need DstSizeTooSmall error which we now have a check for and a test
